// src/constants/enum.ts
export enum Locale {
  en = 'en',
  vi = 'vi'
}

export enum ThemeMode {
  light = 'light',
  dark = 'dark',
  system = 'system'
}

export enum ToastTypes {
  success = 'success',
  error = 'error',
  warning = 'warning',
  info = 'info'
}

export enum Direction {
  vertical = 'vertical',
  horizontal = 'horizontal'
}